chore: split data-table/components/filters/index.tsx
